molgado_writing_suggestions
Davi, congratulations on the work! 

When reading, I felt it is really complete in the technical section (I could visualize the path to use it, even don't really understanding the details). To write something like that is hard, takes time, and didactic effort :sweat-man: 

**But, is it a case or a tutorial?**
The text is mixing a `case` — the process you faced to get to the steps — with a tutorial. Is there a way to make the "why I used this technology" more straight to the point OR explaining more about the problem the tutorial solves, instead of just describing what you faced? 

I thought in one macro suggestion, besides the ones I did on the text: consider writing a little bit more about the *problem* you solved, not just about *how you used Basel and k8s*. Occurs to me that the differential of the content is not *how*, but your sensibility to understand that a monorepo would be a better treatment for the scenario. This may be a way to integrate better "the short case" and "the tutorial" section.

Like, here: 
> we've realized the importance of having a centralized structure where we could know everything that was in production. 
Ok. And what is this importance? Sounds to me an example could be one of these problems that justify a case. If the case doesn't aggregate the tutorial, why keep it? 

Notice that it is preciousness of mine — I don't know if this kind of mixing is common when writing engineering articles. 
The content is already good!